OTel: move OTLP serialization into a partial-class file (reviewability, no behavior change)
